refactor(domain): move status effects into subpackage
