+++ src/main/java/sc2002/turnbased/actions/ShieldBashAction.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import sc2002.turnbased.domain.Combatant;
-import sc2002.turnbased.domain.DamageAdjustment;
-import sc2002.turnbased.domain.StunStatusEffect;
+import sc2002.turnbased.domain.status.DamageAdjustment;
+import sc2002.turnbased.domain.status.StunStatusEffect;
 import sc2002.turnbased.report.ActionEvent;
 import sc2002.turnbased.report.BattleEvent;
