+++ src/main/java/sc2002/turnbased/domain/status/IncomingDamageModifierEffect.java
@@ -1,4 +1,6 @@
-package sc2002.turnbased.domain;
+package sc2002.turnbased.domain.status;
+
+import sc2002.turnbased.domain.Combatant;
 
 public interface IncomingDamageModifierEffect {
     DamageAdjustment adjustIncomingDamage(Combatant owner, Combatant attacker, int damage);
