+++ src/main/java/sc2002/turnbased/domain/status/MergeableStatusEffect.java
@@ -1,4 +1,4 @@
-package sc2002.turnbased.domain;
+package sc2002.turnbased.domain.status;
 
 public interface MergeableStatusEffect {
     boolean canMergeWith(StatusEffect other);
