+++ src/main/java/sc2002/turnbased/domain/status/StatModifierEffect.java
@@ -1,4 +1,6 @@
-package sc2002.turnbased.domain;
+package sc2002.turnbased.domain.status;
+
+import sc2002.turnbased.domain.CombatStats;
 
 public interface StatModifierEffect {
     CombatStats modifyStats(CombatStats stats);
