+++ src/main/java/sc2002/turnbased/domain/status/StunStatusEffect.java
@@ -1,4 +1,4 @@
-package sc2002.turnbased.domain;
+package sc2002.turnbased.domain.status;
 
 import java.util.ArrayList;
 import java.util.List;
