+++ src/main/java/sc2002/turnbased/domain/status/TurnInterferingEffect.java
@@ -1,4 +1,4 @@
-package sc2002.turnbased.domain;
+package sc2002.turnbased.domain.status;
 
 public interface TurnInterferingEffect {
     default TurnEffectResolution onTurnOpportunity() {
